requirements.dev.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-cors-headers==3.9.0
djangorestframework==3.12.4
Faker==8.14.0
idna==3.2
psycopg2==2.9.1
pycodestyle==2.7.0
Pygments==2.10.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
redgreenunittest==0.1.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.7
